♻️ Rename the test modules.
